+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/MockNetwork.kt
@@ -138,6 +138,18 @@ class UnstartedMockNode private constructor(private val node: InternalMockNetwor
      * @return A [StartedMockNode] object.
      */
     fun start(): StartedMockNode = StartedMockNode.create(node.start())
+
+    /**
+     * A [StartedMockNode] object for this running node.
+     * @throws [IllegalStateException] if the node is not running yet.
+     */
+    val started: StartedMockNode
+        get() = StartedMockNode.create(node.started ?: throw IllegalStateException("Node ID=$id is not running"))
+
+    /**
+     * Whether this node has been started yet.
+     */
+    val isStarted: Boolean get() = node.started != null
 }
 
 /** A class that represents a started mock node for testing. */
@@ -421,7 +433,7 @@ open class MockNetwork(
                             forcedID: Int? = null,
                             entropyRoot: BigInteger = BigInteger.valueOf(random63BitValue()),
                             configOverrides: (NodeConfiguration) -> Any? = {},
-                            additionalCordapps: Set<TestCorDapp> = emptySet()): UnstartedMockNode {
+                            additionalCordapps: Set<TestCorDapp>): UnstartedMockNode {
         val parameters = MockNodeParameters(forcedID, legalName, entropyRoot, configOverrides, additionalCordapps)
         return UnstartedMockNode.create(internalMockNetwork.createUnstartedNode(InternalMockNodeParameters(parameters)))
     }